+++ packages/bundle/src/index-core.ts
@@ -3,8 +3,6 @@ import ReactWebChat, {
   Components,
   concatMiddleware,
   connectWithContext,
-  // TODO: [P2] Remove Context
-  Context,
   createStyleSet
 } from 'botframework-webchat-component';
 
@@ -20,7 +18,6 @@ export {
   Components,
   concatMiddleware,
   connectWithContext,
-  Context,
   createBrowserWebSpeechPonyfillFactory,
   createDirectLine,
   createStore,
@@ -32,7 +29,6 @@ window['WebChat'] = {
   ...window['WebChat'],
   concatMiddleware,
   connectWithContext,
-  Context,
   createBrowserWebSpeechPonyfillFactory,
   createDirectLine,
   createStore,
